Rename inngang component files
