+++ src/App.less
@@ -1,6 +1,6 @@
 @import 'components/begrunnelse/begrunnelse.less';
 @import 'components/kvittering/kvittering.less';
-@import 'components/klage-eller-anke/klage-eller-anke.less';
+@import 'components/inngang/inngang.less';
 @import 'pages/oppsummering-skjema-page/oppsummering-skjema-page.less';
 @import 'media-queries.less';
