+++ src/utils/routes.config.tsx
@@ -3,8 +3,8 @@ import FormLandingPage from '../pages/form-landing-page/form-landing-page';
 import OppsummeringSkjemaPage from '../pages/oppsummering-skjema-page/oppsummering-skjema-page';
 import KvitteringPage from '../pages/kvittering/kvittering-page';
 import NotFoundPage from '../pages/not-found/not-found-page';
-import KlageEllerAnkeYtelse from '../components/klage-eller-anke/klage-eller-anke-ytelse';
-import KlageEllerAnkeInnsending from '../components/klage-eller-anke/klage-eller-anke-innsending';
+import KlageEllerAnkeYtelse from '../components/inngang/inngang-kategorier';
+import KlageEllerAnkeInnsending from '../components/inngang/inngang-innsendingsvalg';
 import { TEMA_KEYS } from '../types/tema';
 import { RouteProps } from 'react-router';
 import { INNGANG_KATEGORIER } from '../data/kategorier';
